Use pre-made Procs instead of long case/when construct
Code such as

	case mask_pattern
	when QRMASKPATTERN[:pattern000]
		(i + j) % 2 == 0
	when QRMASKPATTERN[:pattern001]
		i % 2 == 0
	when QRMASKPATTERN[:pattern010]
		j % 3 == 0

is very slow because it requires an `Hash#[]` lookup and a `Kernel#===`
call for every branch of the `case` construct, 3.5 * 2 function calls on
average.

We exploit the fact that `mask_pattern` is a Fixnum and pre-populate an
array of possible computations. This takes exactly one `Array#[]`
lookup and `Proc#call` execution per mask pattern.